refactor(throwShopClosedErr): 1- extract checkIsHoliday calendar api to services 2- extract isNotwithinOpeningTimes
